add prob density function class, rename prob value
