+++ probnode/datatype/__init__.py
@@ -1,1 +1,1 @@
-from .probabilityvalue import *
+from .probability_value import *
